requirements.txt
requests
textblob
fastapi
uvicorn[standard]
google-cloud-firestore
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
requests
python-dotenv